requirements.txt
flask
torch
torchvision
tensorflow
keras
pillow
numpy
